Authentication

+++ src/main/java/br/com/redeSocialPB/services/CommentService.java
@@ -2,7 +2,8 @@ package br.com.redeSocialPB.services;
 
 import br.com.redeSocialPB.entities.Comment;
 import br.com.redeSocialPB.entities.User;
-import br.com.redeSocialPB.exception.UserNotFoundException;
+import br.com.redeSocialPB.exception.UnauthenticatedUserException;
+import br.com.redeSocialPB.exception.UnauthorizedActionException;
 import br.com.redeSocialPB.repositories.CommentRepository;
 import br.com.redeSocialPB.exception.CommentNotFoundException;
 import br.com.redeSocialPB.repositories.UserRepository;
@@ -13,7 +14,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class CommentService {
+public class CommentService {       // Organizar testes unitários para este service
 
     private CommentRepository commentRepository;
     private UserRepository userRepository;
@@ -35,7 +36,7 @@ public class CommentService {
     public Comment createComment(Comment c, String username) {
         User user = userRepository.findByUsername(username);
         if (user == null){
-            throw new UserNotFoundException("Usuário com username " + username +
+            throw new UnauthenticatedUserException("Usuário com username " + username +
                     " não foi autenticado!");
         }
         c.setDataCriacao(LocalDateTime.now());
@@ -49,12 +50,16 @@ public class CommentService {
 
     public void deleteComment(String id, String username) {
         User user = userRepository.findByUsername(username);
+        if (user == null){
+            throw new UnauthenticatedUserException("Usuário com username " + username +
+                    " não foi autenticado!");
+        }
         Optional<Comment> commentOpt = commentRepository.findById(id);
         if(commentOpt.isPresent()) {
             Comment comment = commentOpt.get();
             if(!comment.getUser().getId().equals(user.getId())) {
-                throw new RuntimeException("O usuário " + username +
-                        " não tem permissão para deleetar este comentário!");   // Desenvolver exceção personalizada
+                throw new UnauthorizedActionException("O usuário " + username +
+                        " não tem permissão para deletar este comentário!");
             }
             commentRepository.deleteById(id);
             return;
@@ -63,10 +68,19 @@ public class CommentService {
                 " não foi encontrado para ser deletado!");
     }
 
-    public Comment updateComment(String id, Comment c) {
+    public Comment updateComment(String id, Comment c, String username) {
+        User user = userRepository.findByUsername(username);
+        if (user == null){
+            throw new UnauthenticatedUserException("Usuário com username " + username +
+                    " não foi autenticado!");
+        }
         Optional<Comment> commentOpt = commentRepository.findById(id);
         if(commentOpt.isPresent()) {
             Comment toUpdate = commentOpt.get();
+            if(!toUpdate.getUser().getId().equals(user.getId())) {
+                throw new UnauthorizedActionException("O usuário " + username +
+                        " não tem permissão para atualizar este comentário!");
+            }
             toUpdate.setComentario(c.getComentario());
             toUpdate.setDataAtualizacao(LocalDateTime.now());
             return commentRepository.save(toUpdate);
